+++ resources/views/labels/index.blade.php
@@ -2,8 +2,9 @@
 @section('content')
     <main class="container py-4">
         <h1 class="mb-5">Метки</h1>
-        <a href="{{ route('labels.create') }}" class="btn btn-primary">
-            Создать метку        </a>
+        @auth
+            <a href="{{ route('labels.create') }}" class="btn btn-primary">Создать метку</a>
+        @endauth
         <table class="table mt-2">
             <thead>
             <tr>
@@ -11,7 +12,9 @@
                 <th>Имя</th>
                 <th>Описание</th>
                 <th>Дата создания</th>
-                <th>Действия</th>
+                @auth()
+                    <th>Действия</th>
+                @endauth
             </tr>
             </thead>
             <tbody>
@@ -22,11 +25,14 @@
                     <td>{{ $label->description }}</td>
                     <td>{{ $label->created_at }}</td>
                     <td>
-                        @auth
-                        <a class="text-danger text-decoration-none" href="{{ route('labels.destroy', $label) }}"
-                           data-confirm="Вы уверены?" data-method="delete" rel="nofollow">Удалить</a>
-                        <a class="text-decoration-none" href="{{ route('labels.edit', $label->id) }}">Изменить</a>
-                        @endauth
+                        @can('delete', $label)
+                            <a class="text-danger text-decoration-none" href="{{ route('labels.destroy', $label) }}"
+                               data-confirm="Вы уверены?" data-method="delete" rel="nofollow">Удалить</a>
+                        @endcan
+                            @can('update', $label)
+                                <a class="text-decoration-none"
+                                   href="{{ route('labels.edit', $label->id) }}">Изменить</a>
+                            @endcan
                     </td>
                 </tr>
             @endforeach
